- Re-work calibration and v/oct scaling - CV inputs have calibrated 0V setting (since v/oct input might be configured differently)

+++ braids/drivers/cv_adc.h
@@ -69,7 +69,7 @@ class CvAdc {
     }
   }
 
-  inline uint16_t value(size_t channel) const {
+  inline int32_t value(size_t channel) const {
     return values_[channel];
   }
